+++ src/Libraries/SmartStore.Core/Extensions/EnumerableExtensions.cs
@@ -286,6 +286,11 @@ namespace SmartStore
 			return sorted;
 		}
 
+		public static string StrJoin(this IEnumerable<string> source, string separator)
+		{
+			return string.Join(separator, source);
+		}
+
 		#endregion
 
 		#region Multimap